Add TddTest¥TddTestBase
‘TddTest¥TddTestBase’ is the base class for test.
